chore: enable soft deletes (#2007)

+++ apps/api/src/routes/v1_apis_deleteKeys.ts
@@ -1,6 +1,6 @@
 import type { App } from "@/pkg/hono/app";
 import { createRoute, z } from "@hono/zod-openapi";
-import { eq, sql } from "@unkey/db";
+import { and, eq, isNull, sql } from "@unkey/db";
 
 import { rootKeyAuth } from "@/pkg/auth/root_key";
 import { UnkeyApiError, openApiErrorResponses } from "@/pkg/errors";
@@ -23,10 +23,9 @@ const route = createRoute({
               description: "The id of the api, that the keys belong to.",
               example: "api_1234",
             }),
-            permanent: z.boolean().optional().openapi({
-              deprecated: true,
+            permanent: z.boolean().default(false).optional().openapi({
               description:
-                "Keys are always deleted permanently, this flag is deprecated and will be removed in a future version.",
+                "Delete the keys permanently, if false the keys will be marked as deleted but not removed from the database. In either case, the keys will no longer be valid when verifying them.",
             }),
           }),
         },
@@ -60,7 +59,7 @@ export type V1ApisDeleteKeysResponse = z.infer<
 
 export const registerV1ApisDeleteKeys = (app: App) =>
   app.openapi(route, async (c) => {
-    const { apiId } = c.req.valid("json");
+    const { apiId, permanent } = c.req.valid("json");
     const { cache, db } = c.get("services");
 
     const auth = await rootKeyAuth(
@@ -101,12 +100,27 @@ export const registerV1ApisDeleteKeys = (app: App) =>
     }
 
     let deletedKeys = 0;
-    const where = eq(schema.keys.keyAuthId, api.keyAuthId);
-    await db.primary.transaction(async (tx) => {
-      const keys = await tx.select({ count: sql<string>`count(*)` }).from(schema.keys).where(where);
-      await tx.delete(schema.keys).where(where).execute();
-      deletedKeys = Number.parseInt(keys.at(0)?.count ?? "0");
-    });
+    if (permanent) {
+      const where = eq(schema.keys.keyAuthId, api.keyAuthId);
+      await db.primary.transaction(async (tx) => {
+        const keys = await tx
+          .select({ count: sql<string>`count(*)` })
+          .from(schema.keys)
+          .where(where);
+        await tx.delete(schema.keys).where(where).execute();
+        deletedKeys = Number.parseInt(keys.at(0)?.count ?? "0");
+      });
+    } else {
+      const where = and(eq(schema.keys.keyAuthId, api.keyAuthId), isNull(schema.keys.deletedAt));
+      await db.primary.transaction(async (tx) => {
+        const keys = await tx
+          .select({ count: sql<string>`count(*)` })
+          .from(schema.keys)
+          .where(where);
+        await tx.update(schema.keys).set({ deletedAt: new Date() }).where(where).execute();
+        deletedKeys = Number.parseInt(keys.at(0)?.count ?? "0");
+      });
+    }
 
     return c.json({ deletedKeys });
   });
